requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
certifi==2021.5.30
charset-normalizer==2.0.6
Django==3.2.7
django-ckeditor==6.1.0
django-js-asset==1.2.2
idna==3.2
Pillow==8.3.2
pytz==2021.1
requests==2.26.0
soupsieve==2.2.1
sqlparse==0.4.2
urllib3==1.26.7
